pyproject.toml
[project]
name = "pyorchestrator"
version = "1.0.0"
description = ""
authors = [
    {name = "Kayan Araujo"}
]
license = {text = "Apache License 2.0"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml (>=6.0.2,<7.0.0)",
    "boto3 (>=1.36.16,<2.0.0)",
    "click (>=8.1.8,<9.0.0)",
    "schema (>=0.7.7,<0.8.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
pyorchestrator = 'pyorchestrator.main:cli'
